backend/requirements.txt
celery[redis]==5.4.0
fastapi==0.115.0
firecrawl-py>=1.4.0
google-genai==1.5.0
httpx==0.28.1
numpy==2.1.0
pytest==8.3.5
pytest-asyncio==0.24.0
python-dotenv==1.0.1
supabase==2.28.3
trafilatura==1.12.0
uvicorn==0.30.0
